pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mpm-cudajax"
version = "0.1.0"
requires-python = ">=3.10"
description = "MLS-MPM benchmark: JAX vs PyTorch vs CUDA"
dependencies = [
    "hydra-core>=1.3",
    "omegaconf",
    "matplotlib",
    "tqdm",
    "numpy",
]

[project.optional-dependencies]
jax = ["jax>=0.4.20", "jaxlib>=0.4.20"]
pytorch = ["torch>=2.0.1", "warp-lang>=1.0"]
all = ["mpm-cudajax[jax,pytorch]"]

[tool.setuptools.packages.find]
include = ["mpm_jax*"]

[tool.setuptools.package-dir]
# Make vendored MPM-PyTorch importable
"mpm_pytorch" = "vendor/MPM-PyTorch/mpm_pytorch"

[tool.pytest.ini_options]
testpaths = ["tests"]
